Rename checkForDuplicateIssue to checkForExisting for more clarity
