+++ lib/pull-request-merged-handler.js
@@ -1,4 +1,4 @@
-const checkForDuplicateIssue = require('./utils/check-for-duplicate-issue')
+const checkForExistingIssue = require('./utils/check-for-existing-issue')
 const reopenClosed = require('./utils/reopen-closed')
 const { assignFlow, lineBreak, truncate } = require('./utils/helpers')
 const { issueFromMerge } = require('./templates')
@@ -25,9 +25,9 @@ module.exports = async context => {
     username,
     type
   }) => {
+    const existingIssue = await checkForExistingIssue(context, title)
     if (type === 'add') {
       // Prevent duplicates
-      const existingIssue = await checkForDuplicateIssue(context, title)
       if (existingIssue) {
         context.log(`Duplicate issue found with title [${title}]`)
         return reopenClosed({ context, config, issue: existingIssue }, {
