+++ lib/utils/check-for-existing-issue.js
